+++ VirtualList/tests/VirtualGridList-native-specs.js
@@ -65,7 +65,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -81,7 +80,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -98,7 +96,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -115,7 +112,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -127,7 +123,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -145,7 +140,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				direction="horizontal"
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -165,7 +159,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 					horizontalScrollbar="visible"
 					itemRenderer={renderItem}
 					itemSize={itemSize}
-					scrollMode="native"
 					verticalScrollbar="visible"
 				/>
 			);
@@ -186,7 +179,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 					horizontalScrollbar="hidden"
 					itemRenderer={renderItem}
 					itemSize={itemSize}
-					scrollMode="native"
 					verticalScrollbar="visible"
 				/>
 			);
@@ -209,7 +201,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 					itemRenderer={renderItem}
 					itemSize={itemSize}
 					horizontalScrollbar="hidden"
-					scrollMode="native"
 					verticalScrollbar="hidden"
 				/>
 			);
@@ -240,7 +231,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 					dataSize={itemArray.length}
 					itemRenderer={renderItemArray}
 					itemSize={itemSize}
-					scrollMode="native"
 				/>
 			);
 
@@ -251,7 +241,6 @@ describe('VirtualGridList with native `scrollMode`', () => {
 				dataSize={itemArray.length}
 				itemRenderer={renderItemArray}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>);
 
 			jest.useFakeTimers();