+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The following is a curated list of changes in the Enact agate module, newest changes on the top.
 
+## [unreleased]
+
+Changed `scrollMode` default value as `native` on the following components: `Scroller`, `VirtualList` and `VirtualGridList`.
+
 ## [2.0.6] - 2023-06-08
 
 No significant changes.