+++ CHANGELOG.md
@@ -4,7 +4,9 @@ The following is a curated list of changes in the Enact agate module, newest cha
 
 ## [unreleased]
 
-Changed `scrollMode` default value as `native` on the following components: `Scroller`, `VirtualList` and `VirtualGridList`.
+### Changed
+
+- `agate/Scroller`, `agate/VirtualList` and `agate/VirtualGridList` default value for `scrollMode` to `native`.
 
 ## [2.0.6] - 2023-06-08